im change veriables.

+++ src/Task_1_2_13/Okr.java
@@ -5,7 +5,7 @@ public class Okr {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("square-");
-        double area = in.nextInt();
+        double area = in.nextDouble();
         double ros= calcCircleRadius( area);
         System.out.printf("square-%.0f radius-%.3f",+area,ros);
